requirements.txt
pytorch >= 1.10
geoopt >= 0.4